refactor(deps): move CodeTracking to Revise extension
Access `CodeTracking` via `Revise.CodeTracking` instead of carrying it
as a direct dependency. Moves `_method_offset` and `_has_uuid` into the
Revise extension, with a no-op fallback in the main module. Simplifies
`_source_info` to always dispatch through `ReviseIsLoaded()`.
