+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refine Library component design system: slate neutrals, indigo accents, precise shadows, faster transitions [#51]
+- Move `CodeTracking` from direct dependency to Revise extension via `Revise.CodeTracking` [#58]
 
 ## [v2.2.3] - 2025-06-25
 
@@ -155,3 +156,4 @@ Initial release.
 [#40]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/40
 [#50]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/50
 [#51]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/51
+[#58]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/58
