+++ Project.toml
@@ -4,7 +4,6 @@ authors = ["Michael Hatherly <michaelhatherly@gmail.com>"]
 version = "2.2.3"
 
 [deps]
-CodeTracking = "da1fd8a2-8d9e-5ec2-8556-3022fb5608a2"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 PackageExtensionCompat = "65ce6f38-6b18-4e1d-a461-8949797d7930"
@@ -24,7 +23,6 @@ HypertextTemplatesHTTPExt = "HTTP"
 HypertextTemplatesReviseExt = "Revise"
 
 [compat]
-CodeTracking = "1.3, 2, 3"
 InteractiveUtils = "1.6"
 MacroTools = "0.5"
 PackageExtensionCompat = "1"
